Set amount as int

+++ xrp_api.py
@@ -70,7 +70,7 @@ class XRPAPI():
         return self._call(endpoint=endpoint)
 
     def submit_payment(self, source_address: str, destination_address: str,
-                       amount: int, api_key: str, submit: bool=True) -> dict:
+                       amount: float, api_key: str, submit: bool=True) -> dict:
         '''Sign a payment transaction and submit it to the XRP Ledger network. The
         sending account must match an account address and secret the XRP-API server
         is configured with.
